Refactored all the classes into their own files
