+++ src/DnnSummit/Models/GeneralSponsor.cs
@@ -1,0 +1,8 @@
+﻿using System.Collections.Generic;
+
+namespace DnnSummit.Models
+{
+    public class GeneralSponsor : BaseSponsorGroup<IEnumerable<Sponsor>>
+    {
+    }
+}
