+++ src/DnnSummit/Models/ISponsorGroup.cs
@@ -1,0 +1,7 @@
+﻿namespace DnnSummit.Models
+{
+    public interface ISponsorGroup
+    {
+        SponsorType Level { get; }
+    }
+}
