+++ src/DnnSummit/Models/SponsorType.cs
@@ -1,0 +1,11 @@
+﻿namespace DnnSummit.Models
+{
+    public enum SponsorType
+    {
+        Title,
+        Platinum,
+        Gold,
+        Silver,
+        Bronze
+    }
+}
