+++ src/DnnSummit/Models/TitlePlatinumSponsor.cs
@@ -1,0 +1,6 @@
+﻿namespace DnnSummit.Models
+{
+    public class TitlePlatinumSponsor : BaseSponsorGroup<Sponsor>
+    {
+    }
+}
